prevent unauthorized queries

+++ src/routes/customer/get_time_out/[customer_id].json.js
@@ -3,14 +3,18 @@
 
 import { pool } from '$lib/db';
 
-export const get = async ({ params }) => {
-    try {
+export const get = async (event) => {
+    if (!event.locals.user) {
+        return {
+            status: 401,
+            body: 'Please log in!'
+        }
+    } try {
         // Remove old time-out's
         await pool.query(`
             DELETE FROM time_out_table
             WHERE end_time :: DATE < CURRENT_DATE;
         `);
-        const { customer_id } = params;
         //  Get time-out's
         const res = await pool.query(`
             SELECT
@@ -20,7 +24,7 @@ export const get = async ({ params }) => {
             FROM time_out_table
             WHERE customer_id = $1
             ORDER BY start_date ASC;
-            `, [customer_id]);
+            `, [event.params.customer_id]);
         return {
             body: res.rows
         }
